Issue #233: Move endpoints related to amenities from community tag
