+++ api/src/main/resources/public/swagger/api.yaml
@@ -139,7 +139,7 @@ paths:
       security:
         - bearerAuth: [ ]
       tags:
-        - Communities
+        - Amenities
       description: Get all amenities of community
       operationId: listAllAmenities
       parameters:
@@ -170,7 +170,7 @@ paths:
       security:
         - bearerAuth: [ ]
       tags:
-        - Communities
+        - Amenities
       description: Adds amenity to community
       operationId: addAmenityToCommunity
       requestBody:
