+++ service/src/main/java/com/myhome/controllers/CommunityController.java
@@ -18,14 +18,10 @@ package com.myhome.controllers;
 
 import com.myhome.api.CommunitiesApi;
 import com.myhome.controllers.dto.CommunityDto;
-import com.myhome.controllers.mapper.AmenityApiMapper;
 import com.myhome.controllers.mapper.CommunityApiMapper;
-import com.myhome.domain.Amenity;
 import com.myhome.domain.Community;
 import com.myhome.domain.CommunityHouse;
 import com.myhome.domain.User;
-import com.myhome.model.AddAmenityRequest;
-import com.myhome.model.AddAmenityResponse;
 import com.myhome.model.AddCommunityAdminRequest;
 import com.myhome.model.AddCommunityAdminResponse;
 import com.myhome.model.AddCommunityHouseRequest;
@@ -33,12 +29,10 @@ import com.myhome.model.AddCommunityHouseResponse;
 import com.myhome.model.CommunityHouseName;
 import com.myhome.model.CreateCommunityRequest;
 import com.myhome.model.CreateCommunityResponse;
-import com.myhome.model.GetAmenityDetailsResponse;
 import com.myhome.model.GetCommunityDetailsResponse;
 import com.myhome.model.GetCommunityDetailsResponseCommunity;
 import com.myhome.model.GetHouseDetailsResponse;
 import com.myhome.model.ListCommunityAdminsResponse;
-import com.myhome.services.AmenityService;
 import com.myhome.services.CommunityService;
 import java.util.Arrays;
 import java.util.HashSet;
@@ -65,8 +59,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class CommunityController implements CommunitiesApi {
   private final CommunityService communityService;
   private final CommunityApiMapper communityApiMapper;
-  private final AmenityService amenitySDJpaService;
-  private final AmenityApiMapper amenityApiMapper;
 
   @Override
   public ResponseEntity<CreateCommunityResponse> createCommunity(@Valid @RequestBody
@@ -211,24 +203,5 @@ public class CommunityController implements CommunitiesApi {
     } else {
       return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
-  }
-
-  @Override
-  public ResponseEntity<Set<GetAmenityDetailsResponse>> listAllAmenities(
-      @PathVariable String communityId) {
-    Set<Amenity> amenities = amenitySDJpaService.listAllAmenities(communityId);
-    Set<GetAmenityDetailsResponse> response =
-        amenityApiMapper.amenitiesSetToAmenityDetailsResponseSet(amenities);
-    return ResponseEntity.ok(response);
-  }
-
-  @Override
-  public ResponseEntity<AddAmenityResponse> addAmenityToCommunity(
-      @PathVariable String communityId,
-      @RequestBody AddAmenityRequest request) {
-    return amenitySDJpaService.createAmenities(request.getAmenities(), communityId)
-        .map(amenityList -> new AddAmenityResponse().amenities(amenityList))
-        .map(ResponseEntity::ok)
-        .orElse(ResponseEntity.notFound().build());
   }
 }
